First part of app rewrite in ts (non working)
