+++ index.d.ts
@@ -1,3 +1,4 @@
 declare module '*.png'
 declare module '*.jpg'
-declare module '*.svg';
+declare module '*.svg'
+declare module 'lodash.clonedeep'
