--- src/lib/Utils.js
@@ -1,582 +1,0 @@
-import { IMPORT_NOTE_POSITIONS, APP_NAME, INSTRUMENTS, PITCHES, NOTE_NAMES, LAYOUT_DATA } from "appConfig"
-import LoggerStore from "stores/LoggerStore";
-import * as workerTimers from 'worker-timers';
-class Recording {
-	constructor() {
-		this.start = new Date().getTime()
-		this.notes = []
-	}
-	init = () => {
-		this.start = new Date().getTime() - 100
-		console.log("Started new recording")
-	}
-	addNote = (index) => {
-		if (this.notes.length === 0) this.init()
-		let currentTime = new Date().getTime()
-		let note = [index, currentTime - this.start]
-		this.notes.push(note)
-	}
-}
-
-class Song {
-	constructor(name, notes = [], data = {}) {
-		this.name = name
-		this.version = 1
-		this.notes = notes
-		this.bpm = 220
-		this.pitch = "C"
-		this.data = {
-			isComposed: false,
-			isComposedVersion: false,
-			appName: APP_NAME
-		}
-		Object.entries(data).forEach((entry) => {
-			this.data[entry[0]] = entry[1]
-		})
-	}
-}
-
-
-class NoteData {
-	constructor(index, noteNames, url) {
-		this.index = index
-		this.noteNames = noteNames
-		this.url = url
-		this.buffer = new ArrayBuffer(8)
-		this.clicked = false
-	}
-}
-
-class PlayingSong {
-	constructor(notes) {
-		this.timestamp = new Date().getTime()
-		this.notes = notes
-	}
-}
-function capitalize(str){
-	return str.charAt(0).toUpperCase() + str.slice(1);
-}
-class FileDownloader {
-	constructor(type) {
-		if (type === undefined) type = "text/json"
-		this.dataType = "data:" + type + ";charset=utf-8,"
-	}
-	download = (file, name) => {
-		let data = this.dataType + encodeURIComponent(file)
-		var el = document.createElement("a")
-		el.style = "display:none"
-		document.body.appendChild(el)
-		el.setAttribute("href", data)
-		el.setAttribute("download", name)
-		el.click();
-		el.remove();
-	}
-}
-
-const TempoChangers = [
-	{
-		id: 0,
-		text: "1",
-		changer: 1,
-		color: 0x515c6f
-	}, {
-		id: 1,
-		text: "1/2",
-		changer: 1 / 2,
-		color: 0x517553
-	}, {
-		id: 2,
-		text: "1/4",
-		changer: 1 / 4,
-		color: 0x434c7d
-	}, {
-		id: 3,
-		text: "1/8",
-		changer: 1 / 8,
-		color: 0x774D6D
-	}
-]
-
-class ComposedSong {
-	constructor(name, notes = [], data = {}) {
-		data.isComposed = true
-		data.isComposedVersion = true
-		this.version = 1
-		this.data = data
-		this.name = name
-		this.bpm = 220
-		this.pitch = "C"
-		this.notes = notes
-		this.instruments = [INSTRUMENTS[0], INSTRUMENTS[0], INSTRUMENTS[0]]
-		this.breakpoints = [0]
-		this.columns = []
-		this.selected = 0
-		new Array(100).fill().forEach((e) => {
-			let column = new Column()
-			column.tempoChanger = 0
-			this.columns.push(column)
-		})
-	}
-}
-
-
-class MIDINote{
-	index = 0
-	midi = 0
-	status = 'wrong'
-	constructor(index,midi){
-		this.index = index
-		this.midi = midi
-		this.status = midi < 0 ? 'wrong' : 'right'
-	}
-}
-class MIDIShortcut{
-    constructor(type, midi){
-        this.type = type
-        this.midi = midi
-		this.status = midi < 0 ? 'wrong' : 'right'
-    }
-}
-function getNoteText(noteNameType, index, pitch, layoutLength) {
-    try {
-		const layout = LAYOUT_DATA[layoutLength]
-        if (noteNameType === "Note name") return NOTE_NAMES[APP_NAME][PITCHES.indexOf(pitch)][index]
-        if (noteNameType === "Keyboard layout") return layout.keyboardLayout[index]
-        if (noteNameType === "Do Re Mi") return layout.mobileLayout[index]
-		if (noteNameType === "ABC") return layout.abcLayout[index]
-		if( noteNameType === "No Text") return ''
-    } catch (e) { }
-    return ''
-}
-
-function NotesTable(length){
-	return new Array(length).fill().map(() => {return []})
-}
-
-function ComposerSongDeSerialization(song) {
-	const bpm = Number(song.bpm)
-	const obj = {
-		data: song.data,
-		name: song.name,
-		bpm: isNaN(bpm) ? 220 : bpm,
-		pitch: song.pitch ?? "C",
-		instruments: song.instruments || [INSTRUMENTS[0], INSTRUMENTS[0], INSTRUMENTS[0]],
-		breakpoints: song.breakpoints ?? [],
-		notes: [],
-		selected: 0,
-		columns: []
-	}
-	song.columns.forEach(column => {
-		const columnObj = new Column()
-		columnObj.tempoChanger = column[0]
-		column[1].forEach(note => {
-			columnObj.notes.push(new ColumnNote(note[0], note[1]))
-		})
-		obj.columns.push(columnObj)
-	})
-	return obj
-}
-
-function ComposerToRecording(song) {
-	const recordedSong = new Song(song.name)
-	const bpmPerMs = Math.floor(60000 / song.bpm)
-	let totalTime = 100
-	song.columns.forEach(column => {
-		column[1].forEach(note => {
-			recordedSong.notes.push([note[0], totalTime])
-		})
-		totalTime += Math.floor(bpmPerMs * TempoChangers[column[0]].changer)
-	})
-	return recordedSong
-}
-
-function ComposerSongSerialization(song) {
-	const bpm = Number(song.bpm)
-	const obj = {
-		data: song.data,
-		name: song.name,
-		bpm: isNaN(bpm) ? 220 : bpm,
-		pitch: song.pitch,
-		breakpoints: song.breakpoints,
-		instruments: song.instruments,
-		columns: []
-	}
-	obj.data.appName = APP_NAME
-	/*
-		notes = [tempoChanger,notes] ----> note = [index,layer]
-		tempoChanger = Number
-	*/
-	song.columns.forEach(column => {
-		let columnArr = [column.tempoChanger]
-		let notes = column.notes.map(note => {
-			return [note.index, note.layer]
-		})
-		columnArr[1] = notes
-		obj.columns.push(columnArr)
-	})
-	return obj
-}
-
-function prepareSongDownload(song) {
-	let finalSong = JSON.parse(JSON.stringify(song)) //lose reference
-	let oldFormatNotes = {}
-	if (finalSong.data?.isComposedVersion) {
-		oldFormatNotes = composedToOldFormat(finalSong)
-	} else {
-		oldFormatNotes = recordedToOldFormat(finalSong)
-	}
-	finalSong.isComposed = finalSong.data.isComposedVersion
-	finalSong.pitchLevel = PITCHES.indexOf(finalSong.pitch)
-	finalSong.songNotes = oldFormatNotes
-	finalSong.bitsPerPage = 16
-	finalSong.isEncrypted = false
-	return [finalSong] //for compatibility, add back to an array, it will be ignored when importing
-}
-
-function prepareSongImport(song) {
-	if (Array.isArray(song) && song.length > 0) song = song[0]
-	let type = getSongType(song)
-	if (type === "none") {
-		//TODO maybe not the best place to have these
-		LoggerStore.error("Invalid song")
-		throw new Error("Error Invalid song")
-	}
-	if (type === "oldSky") {
-		song = oldSkyToNewFormat(song)
-	}
-	if (APP_NAME === 'Sky' && song.data?.appName !== 'Sky') {
-		LoggerStore.error("Invalid song")
-		throw new Error("Error Invalid song")
-	}
-	if (APP_NAME === 'Genshin' && song.data?.appName === 'Sky') {
-		song = newSkyFormatToGenshin(song)
-	}
-	return song
-}
-
-function composedToOldFormat(song) {
-	let convertedNotes = []
-	let bpmPerMs = Math.floor(60000 / song.bpm)
-	let totalTime = 100
-	song.columns.forEach(column => {
-		column[1].forEach(note => {
-			let layer = 1
-			if (note[1] === '111') layer = 3
-			if (note[1] === '011') layer = 2
-			if (note[1] === '101') layer = 3
-			if (note[1] === '001') layer = 2
-			if (note[1] === '110') layer = 3
-			if (note[1] === '010') layer = 2
-			if (note[1] === '100') layer = 1
-			let noteObj = {
-				key: (layer > 2 ? 2 : layer) + 'Key' + note[0],
-				time: totalTime
-			}
-			if (layer > 2) noteObj.l = 3
-			convertedNotes.push(noteObj)
-		})
-		totalTime += Math.floor(bpmPerMs * TempoChangers[column[0]].changer)
-	})
-	return convertedNotes
-}
-
-function recordedToOldFormat(song) {
-	let convertedNotes = []
-	song.notes.forEach(note => {
-		convertedNotes.push({
-			time: note[1],
-			key: "1Key" + note[0]
-		})
-	})
-	return convertedNotes
-}
-
-function getSongType(song) {
-	try {
-		if (song.data === undefined) {
-			//oldSky format
-			song.pitchLevel = song.pitchLevel || 0
-			if (song.songNotes !== undefined) {
-				return "oldSky"
-			}
-		} else {
-			//current format
-			if (song.data.isComposedVersion) {
-				if (typeof song.name !== "string") return "none"
-				if (typeof song.bpm !== "number") return "none"
-				if (!PITCHES.includes(song.pitch)) return "none"
-				if (Array.isArray(song.breakpoints)) {
-					if (song.breakpoints.length > 0) {
-						if (typeof song.breakpoints[0] !== "number") return "none"
-					}
-				} else {
-					return "none"
-				}
-				if (Array.isArray(song.columns)) {
-					if (song.columns.length > 0) {
-						let column = song.columns[0]
-						if (typeof column[0] !== "number") return "none"
-					}
-				} else {
-					return "none"
-				}
-				return "newComposed"
-			} else {
-				if (typeof song.name !== "string") return "none"
-				if (typeof song.bpm !== "number") return "none"
-				if (!PITCHES.includes(song.pitch)) return "none"
-				return "newRecorded"
-			}
-		}
-
-	} catch (e) {
-		console.log(e)
-		return "none"
-	}
-	return "none"
-}
-
-
-function newSkyFormatToGenshin(song) {
-	if (song.data.isComposedVersion) {
-		song.instruments = song.instruments.map(instrument => INSTRUMENTS[0])
-		song.columns.forEach(column => {
-			column[1] = column[1].map(note => {
-				return [IMPORT_NOTE_POSITIONS[note[0]], note[1]]
-			})
-
-		})
-	}
-	if (!song.data.isComposedVersion) {
-		song.notes = song.notes.map(note => {
-			note[0] = IMPORT_NOTE_POSITIONS[note[0]]
-			return note
-		})
-	}
-	return song
-}
-
-function oldSkyToNewFormat(song) {
-	let result = new Song("Error")
-	try {
-		const bpm = Number(song.bpm)
-		result = new Song(song.name)
-		result.bpm = isNaN(bpm) ? 220 : bpm
-		result.pitch = (PITCHES[song.pitchLevel || 0]) || "C"
-		//remove duplicate notes
-		song.songNotes = song.songNotes.filter((note, index, self) =>
-			index === self.findIndex((n) => {
-				return n.key.split('Key')[1] === note.key.split('Key')[1] && n.time === note.time
-			})
-		)
-		let songNotes = song.songNotes
-		songNotes.forEach(note => {
-			let data = note.key.split("Key")
-			result.notes.push([IMPORT_NOTE_POSITIONS[data[1]], note.time, note.l ?? Number(data[0])])
-		})
-
-		if ([true, "true"].includes(song.isComposed)) {
-			result = ComposerSongSerialization(RecordingToComposed(result))
-		} else {
-			result.notes = result.notes.map(e => [e[0], e[1]])
-		}
-
-
-	} catch (e) {
-		console.log(e)
-		return new Song("Error importing song, invalid format")
-	}
-	return result
-}
-
-function RecordingToComposed(song,precision = 4) {
-	let bpmToMs = Math.floor(60000 / song.bpm)
-	let composed = new ComposedSong(song.name, [])
-	composed.bpm = song.bpm
-	composed.pitch = song.pitch
-	let notes = song.notes
-	//remove duplicates
-	let converted = []
-	if(precision === 1 ){
-        let groupedNotes = []
-		let previousTime = notes[0][1]
-        while (notes.length > 0) {
-            let row = [notes.shift()]
-            let amount = 0
-            for (let i = 0; i < notes.length; i++) {
-                if (row[0][1] > notes[i][1] - bpmToMs / 9) amount++
-            }
-            groupedNotes.push([...row, ...notes.splice(0, amount)])
-        }
-        let columns = []
-        groupedNotes.forEach(notes => {
-            let note = notes[0]
-            if (!note) return
-            let elapsedTime = note[1] - previousTime
-            previousTime = note[1]
-            let emptyColumns = Math.floor((elapsedTime - bpmToMs) / bpmToMs)
-            if (emptyColumns > -1) new Array(emptyColumns).fill(0).forEach(() => columns.push(new Column())) // adds empty columns
-            let noteColumn = new Column()
-            noteColumn.notes = notes.map(note => {
-                return new ColumnNote(note[0], numberToLayer(note[2] || 0))
-            })
-            columns.push(noteColumn)
-        })
-        columns.forEach(column => { //merges notes of different layer
-            let groupedNotes = groupByIndex(column)
-            column.notes = groupedNotes.map(group => {
-                group[0].layer = mergeLayers(group)
-                return group[0]
-            })
-        })
-		converted = columns
-	}else{
-		let grouped = groupByNotes(notes, bpmToMs / 9)
-		let combinations = [bpmToMs, Math.floor(bpmToMs / 2), Math.floor(bpmToMs / 4), Math.floor(bpmToMs / 8)]
-		for (let i = 0; i < grouped.length; i++) {
-			let column = new Column()
-			column.notes = grouped[i].map(note => {
-				let columnNote = new ColumnNote(note[0])
-				if (note[2] === 0) columnNote.layer = "100"
-				if (note[2] === 1) columnNote.layer = "100"
-				if (note[2] === 2) columnNote.layer = "010"
-				if (note[2] === 3) columnNote.layer = "110"
-				if (note[2] === undefined) columnNote.layer = "100"
-				return columnNote
-			})
-			let next = grouped[i + 1] || [[0, 0, 0]]
-			let difference = next[0][1] - grouped[i][0][1]
-			let paddingColumns = []
-			while (difference >= combinations[3]) {
-				if (difference / combinations[0] >= 1) {
-					difference -= combinations[0]
-					paddingColumns.push(0)
-				} else if (difference / combinations[1] >= 1 ) {
-					difference -= combinations[1]
-					if(precision <= 1) continue
-					paddingColumns.push(1)	
-				} else if (difference / combinations[2] >= 1) {
-					difference -= combinations[2]
-					if(precision <= 2) continue
-					paddingColumns.push(2)
-				} else if (difference / combinations[3] >= 1) {
-					difference -= combinations[3]
-					if(precision <= 3) continue
-					paddingColumns.push(3)
-				}
-			}
-			let finalPadding = []
-			column.tempoChanger = paddingColumns.shift() || 0
-			paddingColumns = paddingColumns.forEach((col, i) => {
-				let column = new Column()
-				column.tempoChanger = col
-				finalPadding.push(column)
-			})
-			converted.push(column, ...finalPadding)
-		}
-	}
-	composed.columns = converted
-	return composed
-}
-
-class Column {
-	constructor() {
-		this.notes = []
-		this.tempoChanger = 0
-
-	}
-}
-
-function groupByNotes(notes, threshold) {
-	let result = []
-	while (notes.length > 0) {
-		let row = [notes.shift()]
-		let amount = 0
-		for (let i = 0; i < notes.length; i++) {
-			if (row[0][1] > notes[i][1] - threshold) amount++
-		}
-		result.push([...row, ...notes.splice(0, amount)])
-	}
-	return result
-}
-
-class ColumnNote {
-	constructor(index, layer = "000") {
-		this.index = index
-		this.layer = layer
-	}
-}
-
-function getPitchChanger(pitch) {
-	let index = PITCHES.indexOf(pitch)
-	if (index < 0) index = 0
-	return Number(Math.pow(2, index / 12).toFixed(4))
-}
-
-function numberToLayer(number) {
-	let layer = "100"
-	if (number === 0) layer = "100"
-	if (number === 1) layer = "010"
-	if (number === 2) layer = "001"
-	return layer
-}
-
-function mergeLayers(notes) {
-	let final = "000".split("")
-	notes.forEach(notes => {
-		notes.layer.split("").forEach((e, i) => {
-			if (e === "1") final[i] = "1"
-		})
-	})
-	return final.join("")
-}
-
-function groupByIndex(column) {
-	let notes = []
-	column.notes.forEach(note => {
-		if (notes[note.index]) {
-			notes[note.index].push(note)
-		} else {
-			notes[note.index] = [note]
-		}
-	})
-	return notes.filter(e => Array.isArray(e))
-}
-
-function delayMs(ms) {
-    return new Promise(resolve => {
-        workerTimers.setTimeout(resolve, ms)
-    })
-}
-
-export {
-	Recording,
-	Song,
-	NoteData,
-	FileDownloader,
-	PlayingSong,
-	ComposedSong,
-	ColumnNote,
-	Column,
-	TempoChangers,
-	ComposerSongSerialization,
-	ComposerSongDeSerialization,
-	ComposerToRecording,
-	getPitchChanger,
-	getSongType,
-	oldSkyToNewFormat,
-	RecordingToComposed,
-	prepareSongDownload,
-	newSkyFormatToGenshin,
-	prepareSongImport,
-	groupByNotes,
-	numberToLayer,
-	mergeLayers,
-	groupByIndex,
-	delayMs,
-	NotesTable,
-	MIDINote,
-	getNoteText,
-	MIDIShortcut,
-	capitalize
-}
